completed jitting for unidir2

+++ Simulator/python/src/Interpolators/MEMCI/ME/tss.py
@@ -10,8 +10,8 @@ from numba import njit, uint32, float32, int8, int32, uint8, int64, types, confi
 @njit(uint32(uint8[:,:,:], uint8[:,:,:]), cache=True)
 def get_sad(source_block, target_block):
     # we need to chagne it to int8 so that it's correct
-    source_block = source_block.astype(np.int8)
-    target_block = target_block.astype(np.int8)
+    source_block = source_block.astype(np.int32)
+    target_block = target_block.astype(np.int32)
     return np.sum(np.abs(np.subtract(source_block, target_block)))
 
 @njit(float32[:,:,:](int32, int32, types.UniTuple(int32, 3), uint8[:,:,:], uint8[:,:,:]), cache=True)
